Add build-inline-extensions build script
