+++ browser/scripts/tasks.ts
@@ -1,6 +1,5 @@
 /* eslint no-sync: warn */
 import fs from 'fs'
-import os from 'os'
 import { omit } from 'lodash'
 import path from 'path'
 import shelljs from 'shelljs'
@@ -63,32 +62,6 @@ function copyExtensionAssets(toDirectory: string): void {
     shelljs.cp('build/dist/options.html', toDirectory)
 }
 
-function copyInlineExtensions(toDirectory: string): void {
-    const extensionName = 'template'
-
-    const temporaryCloneDirectory = path.join(os.tmpdir(), 'code-intel-extensions')
-    shelljs.mkdir('-p', temporaryCloneDirectory)
-
-    // Clone and build
-    shelljs.pushd(temporaryCloneDirectory)
-    shelljs.exec('git clone git@github.com:sourcegraph/code-intel-extensions || (cd code-intel-extensions; git pull)')
-    shelljs.exec('yarn --cwd code-intel-extensions install')
-    shelljs.exec(`yarn --cwd code-intel-extensions/extensions/${extensionName} run build`)
-    shelljs.popd()
-
-    // Copy extension manifest (package.json) and bundle (extension.js)
-    shelljs.mkdir('-p', `${toDirectory}/extensions/${extensionName}`)
-
-    shelljs.cp(
-        `${temporaryCloneDirectory}/code-intel-extensions/extensions/${extensionName}/dist/extension.js`,
-        `${toDirectory}/extensions/${extensionName}`
-    )
-    shelljs.cp(
-        `${temporaryCloneDirectory}/code-intel-extensions/extensions/${extensionName}/package.json`,
-        `${toDirectory}/extensions/${extensionName}`
-    )
-}
-
 export function copyIntegrationAssets(): void {
     shelljs.mkdir('-p', 'build/integration/scripts')
     shelljs.mkdir('-p', 'build/integration/css')
@@ -137,6 +110,8 @@ function writeManifest(environment: BuildEnv, browser: Browser, writeDirectory: 
 
     if (browser === 'firefox') {
         manifest.permissions!.push('<all_urls>')
+        manifest.web_accessible_resources = manifest.web_accessible_resources || []
+        manifest.web_accessible_resources.push('extensions/*')
         delete manifest.storage
     }
 
@@ -169,7 +144,6 @@ function buildForBrowser(browser: Browser): (env: BuildEnv) => () => void {
             signale.await(`Building the ${title} ${environment} bundle`)
 
             copyExtensionAssets(buildDirectory)
-            copyInlineExtensions(buildDirectory)
 
             const zipDestination = path.resolve(process.cwd(), `${BUILDS_DIR}/bundles/${BROWSER_BUNDLE_ZIPS[browser]}`)
             if (zipDestination) {
